requirements-backend.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
surrealdb>=1.0.0
pydantic>=2.4.2
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-anthropic>=0.0.1
langchain-google-genai>=0.0.5
google-cloud-aiplatform>=1.36.0
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
python-docx>=1.0.0
PyPDF2>=3.0.0
epub-conversion>=0.0.3
yt-dlp>=2023.11.16
moviepy>=1.0.3
pydub>=0.25.1
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'
